Added pause menu
Added PauseMenuState.
Added pause and resume methods in the State class.
Added pause and resume methods in the GameStateManager. These can be
called when the ApplicationListener goes into the pause and resume
methods.

At this time if the PlayState goes into a pause it creates a new
PauseMenuState. When PauseMenuState goes into resume it pops this state
off of the stack and resumes the state underneath; the PlayState.