Fixed emission glitch at restart ball position
Created a renderTrail boolean in ParticleManager to handle the state of
rendering of the trail particle effect.
Added a doRenderTrail boolean method to get that state.
Added a startParticleEffect() method to particle manager to handle
particle effect start() method. This should be called instead of the
regular start() method for a particle.